inline superclass
